Move code into simulation.h
